move BrightnessService to utils folder
